Conflicts:
	odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/dao/onto/OntologyDao.java
	odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/pages/transformers/dn/DNGroupDetailPage.java

+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/bo/oi/OIOutput.java
@@ -1,5 +1,7 @@
 package cz.cuni.mff.odcleanstore.webfrontend.bo.oi;
 
+import java.math.BigDecimal;
+
 import cz.cuni.mff.odcleanstore.webfrontend.bo.EntityWithSurrogateKey;
 
 /**
@@ -13,8 +15,8 @@ public class OIOutput extends EntityWithSurrogateKey
 
 	private Integer ruleId;
 	private Integer outputTypeId;
-	private Double minConfidence;
-	private Double maxConfidence;
+	private BigDecimal minConfidence;
+	private BigDecimal maxConfidence;
 	private String filename;
 	private OIFileFormat fileFormat;
 	
@@ -28,8 +30,8 @@ public class OIOutput extends EntityWithSurrogateKey
 	 * @param filename
 	 * @param fileFormat
 	 */
-	public OIOutput(Integer id, Integer ruleId, Integer outputTypeId, Double minConfidence, 
-		Double maxConfidence, String filename, OIFileFormat fileFormat) 
+	public OIOutput(Integer id, Integer ruleId, Integer outputTypeId, BigDecimal minConfidence, 
+		BigDecimal maxConfidence, String filename, OIFileFormat fileFormat) 
 	{
 		super(id);
 		
@@ -70,7 +72,7 @@ public class OIOutput extends EntityWithSurrogateKey
 	 * 
 	 * @return
 	 */
-	public Double getMinConfidence() 
+	public BigDecimal getMinConfidence() 
 	{
 		return minConfidence;
 	}
@@ -79,7 +81,7 @@ public class OIOutput extends EntityWithSurrogateKey
 	 * 
 	 * @return
 	 */
-	public Double getMaxConfidence() 
+	public BigDecimal getMaxConfidence() 
 	{
 		return maxConfidence;
 	}
